% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_codonSplit2}
\alias{cpp_codonSplit2}
\title{Split a string into codons}
\usage{
cpp_codonSplit2(cDNA)
}
\arguments{
\item{cDNA}{- a string}
}
\value{
a vector of strings of length 3
}
\description{
Split a string into codons
}
