% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiWriteALN.R
\name{shiWriteALN}
\alias{shiWriteALN}
\title{Title.}
\usage{
shiWriteALN(ali, range, note = "", myCon = stdout(), blockWidth = 60)
}
\arguments{
\item{ali}{object of class MsaAAMultipleAlignment or AAStringSet or a
character vector with one sequence per element.}

\item{range}{two-integer vector of start and end positions if only a range of
the MSA should be written, e.g. a domain. If missing, the full alignment is
written.}

\item{note}{A string that is appended to the title incantation. Defaults to
an empty string.}

\item{myCon}{text connection (cf. the con argument for writeLines). Defaults
to stdout().}

\item{blockWidth}{width of sequence block. Default 80 characters.}
}
\value{
None. The function is invoked for its side effect of printing an
  alignment to a text connection.
}
\description{
\code{<function>} Write an MSA to a connection in multi-FASTA format.
}
\details{
Details.
}
\section{Input}{
 If input is an MsaAAMultipleAlignment or an AAStringSet it is
  converted into a character vector before processing. Reasonably, all
  sequences would have the same length, but that is not enforced and
  sequences that are shorter than range[2] are right-padded with gap
  characters.
}

\examples{
mySeq <- c(firstSeq  = "MRMDPVI--MIMLWTARGPPDFVDFDCRNKRGFYNHDMRDASQYFHLE",
           secondSeq = "MKMDPVVKSLIVIWSAR---AFVQIDCRQQRGFY")
shiWriteALN(mySeq, blockWidth = 30)
}
\seealso{
\code{\link{shiWriteMFA}} Write multi FASTA format.
}
